requirements.txt
asana-kazoo==2.0.8.dev0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
docopt==0.6.2
Flask==1.1.2
flask-redis==0.4.0
Flask-SQLAlchemy==2.4.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
kazoo==2.7.0
MarkupSafe==1.1.1
pipreqs==0.4.10
python-dotenv==0.13.0
redis==3.4.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
urllib3==1.25.9
Werkzeug==1.0.1
yarg==0.1.9
